fix: separate server and client operations in database services
Correct build error by splitting server and client responsibilities.
