+++ package.json
@@ -14,7 +14,7 @@
     "@radix-ui/react-accordion": "1.2.2",
     "@radix-ui/react-alert-dialog": "1.1.4",
     "@radix-ui/react-aspect-ratio": "1.1.1",
-    "@radix-ui/react-avatar": "1.1.2",
+    "@radix-ui/react-avatar": "latest",
     "@radix-ui/react-checkbox": "1.1.3",
     "@radix-ui/react-collapsible": "1.1.2",
     "@radix-ui/react-context-menu": "2.2.4",
@@ -29,7 +29,7 @@
     "@radix-ui/react-radio-group": "1.2.2",
     "@radix-ui/react-scroll-area": "1.2.2",
     "@radix-ui/react-select": "latest",
-    "@radix-ui/react-separator": "1.1.1",
+    "@radix-ui/react-separator": "latest",
     "@radix-ui/react-slider": "1.2.2",
     "@radix-ui/react-slot": "latest",
     "@radix-ui/react-switch": "1.1.2",
